Analog Calculator using OP AMP
The purpose of our project is to construct an analog calculator with the use of op-amp 741. We 
have enabled this analog calculator to carry out 7 types of linear and non-linear operations: linear 
operations: Addition, Subtraction, Differentiation, Integration, and non-linear operations: 
Logarithmic operation, Exponentiation, Multiplication. We have constructed a Proteus circuit in 
which a user will be able to choose which operation they want the calculator to perform. Then 
they could select the input as a pulse wave, triangular wave, sinusoidal wave, saw tooth wave, or 
DC voltage. They also can vary the frequencies and values from the signal generator as required. 
We can study output without difficulty from the virtual oscilloscope linked to the specific 
channels.