Song Recognition Using Audio Fingerprinting
We have developed a flexible audio search engine. The algorithm is noise and 
distortion resistant, computationally efficient, and capable of quickly identifying a 
short segment of music captured through a microphone in the presence of foreground 
voices and other dominant noise, and through voice codec compression, out of a 
database of many tracks. The algorithm uses a hashed time-frequency constellation 
analysis of the audio. Furthermore, for applications such as radio monitoring, search 
times on the order of a few milliseconds per query are attained, even on a massive 
music database.